+++ build.gradle
@@ -68,7 +68,7 @@ dependencies {
 }
 
 shadowJar {
-    archiveFileName = 'Medimate.jar'
+    archiveFileName = 'Medimate-v1.3.jar'
 }
 
 run {
