+++ docs/diagrams/FilesClassDiagram.puml
@@ -1,0 +1,39 @@
+@startuml
+!include style.puml
+skinparam arrowThickness 1.1
+skinparam arrowColor FILES_COLOR_T4
+skinparam classBackgroundColor FILES_COLOR
+
+package Files <<Rectangle>> {
+  Class "<<interface>>\nFiles" as Files
+  Class "<<interface>>\nFileReader" as FileReader
+  Class "<<interface>>\nFileGenerator" as FileGenerator
+  Class FileManager
+  Class FileStorage
+  Class FileReaderManager
+  Class PdfReader
+  Class ImageReader
+  Class PdfGenerator
+}
+
+Class HiddenOutside #FFFFFF
+HiddenOutside ..> Files
+PdfReader .up.|> FileReader
+PdfGenerator .up.|> FileGenerator
+ImageReader .up.|> FileReader
+FileManager .up.|> Files
+
+FileManager -down-> FileReaderManager : uses
+FileManager -down-> PdfGenerator : uses
+FileManager -down-> FileStorage : uses
+FileReaderManager -down-> PdfReader : uses
+FileReaderManager -down-> ImageReader : uses
+
+package UI {
+Class HiddenModel #FFFFFF
+}
+
+PdfReader --> UI
+ImageReader --> UI
+
+@enduml
