+++ docs/diagrams/style.puml
@@ -31,6 +31,12 @@
 !define STORAGE_COLOR_T3 #806600
 !define STORAGE_COLOR_T2 #544400
 
+!define FILES_COLOR #FFA500
+!define FILES_COLOR_T1 #FFE6CC
+!define FILES_COLOR_T2 #FFD9B3
+!define FILES_COLOR_T3 #FFC680
+!define FILES_COLOR_T4 #FFB347
+
 !define USER_COLOR #000000
 
 skinparam BackgroundColor #FFFFFFF
